Sequential Slideshow mode + Image Caching + Next/Previous Image Controls (#84)
* caching + slideshow controls + non-random order
* explain service state
* delete currupted image seperately!
* memoryRemember explained
* fix duplicate function: select album/image in base
* getImagesFor explanation
* explainState + service remembers current state
* move colormatched image to /tmp
* track #images for each album/service
* micro fix: don't skip image after config change
* fixes for unstable Internet + invalid urls

+++ modules/helper.py
@@ -17,13 +17,41 @@ import subprocess
 import socket
 import logging
 import os
+import shutil
 import re
+import random
+
+# A regular expression to determine whether a url is valid or not (e.g. "www.example.de/someImg.jpg" is missing "http://")
+VALID_URL_REGEX = re.compile(
+	r'^(?:http|ftp)s?://'  # http:// or https://
+	r'(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}\.?|[A-Z0-9-]{2,}\.?)|'  # domain...
+	r'localhost|'  # localhost...
+	r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})'  # ...or ip
+	r'(?::\d+)?'  # optional port
+	r'(?:/?|[/?]\S+)$', re.IGNORECASE)
 
 class helper:
 	@staticmethod
+	def isValidUrl(url):
+		# Catches most invalid URLs
+		if re.match(VALID_URL_REGEX, url) is None:
+			return False
+		return True
+
+	@staticmethod
+	def getWeightedRandomIndex(weights):
+		totalWeights = sum(weights)
+		normWeights = [float(w)/totalWeights for w in weights]
+		x = random.SystemRandom().random()
+		for i in range(len(normWeights)):
+			x -= normWeights[i]
+			if x <= 0.:
+				return i
+
+	@staticmethod
 	def getResolution():
 		res = None
-		output = subprocess.check_output(['/bin/fbset'], stderr=DEVNULL)
+		output = subprocess.check_output(['/bin/fbset'], stderr=subprocess.DEVNULL)
 		for line in output.split('\n'):
 			line = line.strip()
 			if line.startswith('mode "'):
@@ -58,28 +86,82 @@ class helper:
 		return None
 
 	@staticmethod
-	def makeFullframe(filename, displayWidth, displayHeight, zoomOnly=False, autoChoose=False):
-		name, ext = os.path.splitext(filename)
-		filename_temp = "%s-frame%s" % (name, ext)
+	def getMimeType(filename):
+		if not os.path.isfile(filename):
+			return None
+
+		cmd = ["/usr/bin/file", "--mime", filename]
+		with open(os.devnull, 'wb') as void:
+			try:
+				output = subprocess.check_output(cmd, stderr=void).strip("\n")
+				mimetype = output.lstrip(filename+":").strip()
+				if "; charset=" in mimetype:
+					mimetype = mimetype.split("; charset=")[0]
+			except subprocess.CalledProcessError:
+				logging.debug("unable to determine mimetype of file: %s" % filename)
+				return None
+		return mimetype
+
+	@staticmethod
+	def copyFile(orgFilename, newFilename):
+		try:
+			shutil.copyfile(orgFilename, newFilename)
+		except:
+			logging.exception('Unable copy file from "%s" to "%s"'%(orgFilename, newFilename))
+			return False
+		return True
+
+	@staticmethod
+	def scaleImage(orgFilename, newFilename, newSize):
+		cmd = [
+			'convert',
+			orgFilename + '[0]',
+			'-scale',
+			'%sx%s' % (newSize["width"], newSize["height"]),
+			'+repage',
+			newFilename
+        ]
+
+		try:
+			subprocess.check_output(cmd, stderr=subprocess.STDOUT)
+		except subprocess.CalledProcessError as e:
+			logging.exception('Unable to reframe the image')
+			logging.error('Output: %s' % repr(e.output))
+			return False
+		return True
+		
+
+	@staticmethod
+	def getImageSize(filename):
+		if not os.path.isfile(filename):
+			return None
 
 		with open(os.devnull, 'wb') as void:
 			try:
 				output = subprocess.check_output(['/usr/bin/identify', filename], stderr=void)
 			except:
-				logging.exception('Error trying to identify image')
-				return False
+				return None
 
 		m = re.search('([1-9][0-9]*)x([1-9][0-9]*)', output)
 		if m is None or m.groups() is None or len(m.groups()) != 2:
 			logging.error('Unable to resolve regular expression for image size')
-			return False
-		width = int(m.group(1))
-		height = int(m.group(2))
+			return None
+
+		imageSize = {}
+		imageSize["width"] = int(m.group(1))
+		imageSize["height"] = int(m.group(2))
+
+		return imageSize
+
+	@staticmethod
+	def makeFullframe(filename, displayWidth, displayHeight, zoomOnly=False, autoChoose=False):
+		imageSize = helper.getImageSize(filename)
+		width = imageSize["width"]
+		height = imageSize["height"]
 
 		width_border = 15
 		width_spacing = 3
 		border = None
-		borderSmall = None
 		spacing = None
 
 		# Calculate actual size of image based on display
@@ -111,11 +193,11 @@ class helper:
 			else:
 				resizeString = '%sx%s'
 				logging.debug('Image is fullscreen, no reframing needed')
-				return False
+				return filename
 
 			if padding < 20 and not autoChoose:
 				logging.debug('That\'s less than 20px so skip reframing (%dx%d => %dx%d)', width, height, adjWidth, adjHeight)
-				return False
+				return filename
 
 			if padding < 60 and autoChoose:
 				zoomOnly = True
@@ -134,6 +216,12 @@ class helper:
 		try:
 			# Time to process
 			if zoomOnly:
+				p, f = os.path.split(filename)
+				filenameProcessed = os.path.join(p, "zoomed", f)
+				if helper.getImageSize(filenameProcessed):
+					logging.debug("using cached processed image: %s" % filenameProcessed)
+					return filenameProcessed
+
 				cmd = [
 					'convert',
 					filename + '[0]',
@@ -144,9 +232,15 @@ class helper:
 					'-crop',
 					'%sx%s+0+0' % (displayWidth, displayHeight),
 					'+repage',
-					filename_temp
+					filenameProcessed
 				]
 			else:
+				p, f = os.path.split(filename)
+				filenameProcessed = os.path.join(p, "blurred", f)
+				if helper.getImageSize(filenameProcessed):
+					logging.debug("using cached processed image: %s"%filenameProcessed)
+					return filenameProcessed
+
 				cmd = [
 					'convert',
 					filename + '[0]',
@@ -181,24 +275,23 @@ class helper:
 					'%sx%s' % (displayWidth, displayHeight),
 					')',
 					'-composite',
-					filename_temp
+					filenameProcessed
 				]
 		except:
 			logging.exception('Error building command line')
 			logging.debug('Filename: ' + repr(filename))
-			logging.debug('Filename_temp: ' + repr(filename_temp))
+			logging.debug('filenameProcessed: ' + repr(filenameProcessed))
 			logging.debug('border: ' + repr(border))
 			logging.debug('spacing: ' + repr(spacing))
-			return False
+			return None
 
 		try:
 			subprocess.check_output(cmd, stderr=subprocess.STDOUT)
 		except subprocess.CalledProcessError as e:
 			logging.exception('Unable to reframe the image')
 			logging.error('Output: %s' % repr(e.output))
-			return False
-		os.rename(filename_temp, filename)
-		return True
+			return None
+		return filenameProcessed
 
 	@staticmethod
 	def timezoneList():
